+++ src/main/java/blackjack/controller/BlackJackGameController.java
@@ -2,12 +2,12 @@ package blackjack.controller;
 
 import java.util.Map;
 
-import blackjack.domain.Dealer;
-import blackjack.domain.Deck;
-import blackjack.domain.Player;
-import blackjack.domain.Players;
-import blackjack.domain.Result;
-import blackjack.domain.ResultType;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.card.Deck;
+import blackjack.domain.user.Player;
+import blackjack.domain.user.Players;
+import blackjack.domain.result.Result;
+import blackjack.domain.result.ResultType;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
