+++ src/main/java/blackjack/domain/card/Number.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Number {
 	ACE(1),
