+++ src/main/java/blackjack/domain/card/Type.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Type {
 	HEART("하트"),
