+++ src/main/java/blackjack/domain/result/Result.java
@@ -1,8 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Player;
 
 public class Result {
