+++ src/main/java/blackjack/domain/result/ResultType.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.result;
+
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Player;
 
 public enum ResultType {
 	WIN("승"),
