+++ src/main/java/blackjack/domain/user/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import java.util.Objects;
