+++ src/main/java/blackjack/domain/user/Score.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import java.util.HashMap;
 import java.util.Map;
