+++ src/test/java/blackjack/domain/DealerTest.java
@@ -4,6 +4,12 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Number;
+import blackjack.domain.card.Type;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Player;
+
 public class DealerTest {
 	@Test
 	void dealer_score_higher_than_player_score() {
