+++ src/test/java/blackjack/domain/DeckTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+
 public class DeckTest {
 	@Test
 	void distribute_Card() {
