+++ src/test/java/blackjack/domain/NameTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
+import blackjack.domain.user.Name;
+
 public class NameTest {
 	@ParameterizedTest
 	@NullAndEmptySource()
