+++ src/test/java/blackjack/domain/PlayersTest.java
@@ -8,6 +8,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.user.Players;
+
 public class PlayersTest {
 	@Test
 	@DisplayName("빈 리스트를 받았을 때 에러 발생")
